// src/app/modules/dashboard/constants/chart-type.enum.ts
export enum ChartType {
 LINE_CHART = 0,

 PIE_CHART_AVG_TOP5 = 14,
 PIE_CHART_LAST_TOP5 = 13,
 PIE_CHART_LAST_TOP10 = 22,
 PIE_CHART_AVG_TOP10 = 23,
 PIE_CHART_LAST_ALL = 24,
 PIE_CHART_AVG_ALL = 25,
 PIE_CHART_LAST_BOTTOM10 = 26,
 PIE_CHART_AVG_BOTTOM10 = 27,

 BAR_CHART_AVG_ALL = 28,
 BAR_CHART_AVG_TOP5 = 29,
 BAR_CHART_AVG_TOP10 = 30,
 BAR_CHART_AVG_BOTTOM10 = 31,

 AREA_CHART = 6,
 DUAL_LINE_BAR = 4,

 STACKED_AREA_CHART = 12,
 STACKED_BAR_CHART = 11,

 PERCENTILE_GRAPH = 7,
 SLAB_COUNT_GRAPH = 8,

 DIAL_CHART_LAST = 18,
 DIAL_CHART_AVG = 17,

 METER_CHART_LAST = 20,
 METER_CHART_AVG = 19,

 DUAL_AXIS_LINE = 3,

 DONUT_CHART_LAST = 32,
 DONUT_CHART_AVG = 33,

 GEO_MAP_AVG = 16,
 GEO_MAP_LAST = 15,

 LINE_STACKED = 5,
 MIN_MAX_GRAPH = 21,
}